services/gp5-writer/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyguitarpro==0.10.0
